requirements.txt
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.1
passlib==1.7.4
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
